requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
compreface-sdk==0.6.0
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==1.26.15
